requirements.txt
# Counter-Narrative Generator Dependencies

# Vector Store
chromadb>=0.4.22

# OpenRouter API (OpenAI-compatible)
openai>=1.12.0

# Data processing
pandas>=2.0.0
tqdm>=4.66.0

# CLI
rich>=13.7.0
typer>=0.9.0

# Environment
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0
